+++ lib/weber/i18n/i18n.ex
@@ -6,22 +6,21 @@ defmodule Weber.I18n do
   @doc """
   Localize UTC time now
   """
-  def localize_time_now_utc do
-    {date_time_format, locale_process_name} = localize_time_helper
+  def localize_time_now_utc(conn) do
+    {date_time_format, locale_process_name} = localize_time_helper(conn)
     format_time(date_time_format, locale_process_name)
   end
 
-  def localize_time_now(datetime = {{_year, _month, _day}, {_hour, _minute, _second}}) do
-    {date_time_format, locale_process_name} = localize_time_helper
+  def localize_time_now(conn, datetime = {{_year, _month, _day}, {_hour, _minute, _second}}) do
+    {date_time_format, locale_process_name} = localize_time_helper(conn)
     format_time(date_time_format, locale_process_name, datetime)
   end
 
-  defp localize_time_helper do
-    pid = case get_session(:locale) do
+  defp localize_time_helper(conn) do
+    pid = case get_session(conn, :locale) do
       [] -> "en_US.json"
       l -> l <> ".json"
     end
-
     date_time_format = :gen_server.call(binary_to_atom(pid), :get_date_time_format)
     {date_time_format, binary_to_atom(pid)}
   end
@@ -29,8 +28,8 @@ defmodule Weber.I18n do
   @doc """
   Translation helper.
   """
-  def t(key) do
-    case Weber.Session.get_session(:locale) do
+  def t(conn, key) do
+    case Weber.Session.get_session(conn, :locale) do
       [] -> 
         :gen_server.call(:en_US, {:translate, key})
       locale ->