+++ lib/weber/i18n/localization_manager.ex
@@ -14,12 +14,11 @@ defmodule Weber.Localization.LocalizationManager do
     :gen_server.cast(:erlang.self(), :load_localization_files)
     { :ok, LocalizationConfig.new config: config}
   end
-
+  
   def handle_cast(:load_localization_files, state) do
     case :lists.keyfind(:localization, 1, state.config) do
       false ->  
         {:stop, :normal, state}
-
       {:localization, localization_config} ->
         {:ok, project_path} = File.cwd()
         {_, default_locale}  = :lists.keyfind(:default_locale, 1, localization_config)