+++ lib/weber/templates/views_loader.ex
@@ -4,11 +4,11 @@ defmodule Weber.Templates.ViewsLoader do
 
   defmacro set_up_resources(path) do
     path = Code.eval_quoted(path, [], file: __ENV__.file, line: __ENV__.line) |> elem(0)
-    views_list = Enum.filter(get_all_files(:erlang.binary_to_list(path) ++ '/lib/views/'), 
+    views_list_tmp = Enum.filter(get_all_files(:erlang.binary_to_list(path) ++ '/lib/views/'), 
       fn(f) -> :filename.extension(f) == '.html' 
     end)
+    views_list = Enum.map(views_list_tmp, fn(view) -> {Weber.Utils.build_module_name(view), view} end)
     static_list = get_all_files(:erlang.binary_to_list(path) ++ '/public/')
-
     quote do
       defmodule unquote(Weber.Path) do
         def __root__, do: unquote(path)
@@ -21,7 +21,6 @@ defmodule Weber.Templates.ViewsLoader do
   defmacro compile_views(root) do
     root = Code.eval_quoted(root, [], file: __ENV__.file, line: __ENV__.line) |> elem(0)
     views = Enum.filter(get_all_files(:erlang.binary_to_list(root) ++ '/lib/views/'), fn(f) -> :filename.extension(f) == '.html' end)
-  
     lc view inlist views do
       content = add_helpers_imports(File.read!(view))
       quote do
@@ -29,7 +28,7 @@ defmodule Weber.Templates.ViewsLoader do
           require EEx
  
           EEx.function_from_string(:def, :render_template, unquote(content), [:assigns])
-
+          
           def __view__, do: unquote(content)
 
         end